feat(events): add output transmit, checkValid, and new validator
Event transmit is used by the parent component to fetch data.
The event is triggered by the checkValid function.
Unlike the same function in checklist component, checkValid function
here validates the form first and emits the transmit event only when
the form is valid.
New validators are introduced to check if date range is valid and
if the required input field in the selected case is filled or not.
More details can be found in the two components file.