fix(events): fix bugs caused by date and frequency and add a new field
Fix the bug that formDataToEvent can't correctly convert start date and
end date. It incorrectly extracted date from form data and has been
corrected to extract date from schedule in the form data.

Fix the bug that event can't store repeat frequency in nested arrays.
Change the repeat frequency to an object array.

Add another field groupip to Event. If the event is not a group event,
then this field will be blank.